Crate api_server

Implements the interface for intercepting API requests, forwarding them to the VMM and responding to the user. It is constructed on top of an HTTP Server that uses Unix Domain Sockets and EPOLL to handle multiple connections on the same thread.

Modules

parsed_request 🔒

request 🔒

Structs

ApiServer

Structure associated with the API server implementation.

Body

The Body associated with an HTTP Request or Response.

HttpServer

HTTP Server implementation using Unix Domain Sockets and EPOLL to handle multiple connections on the same thread.

Request

Wrapper over an HTTP Request.

Response

Wrapper over an HTTP Response.

ServerRequest

Wrapper over Request which adds an identification token.

ServerResponse

Wrapper over Response which adds an identification token.

Enums

Method

Supported HTTP Methods.

RequestError

Errors associated with parsing the HTTP Request from a u8 slice.

ServerError

Errors pertaining to HttpServer.

StatusCode

Wrapper over a response status code.

Version

Supported HTTP Versions.