Module api_server::request

Modules

actions

balloon

boot_source

cpu_configuration

drive

entropy

instance_info

logger

machine_configuration

metrics

mmds

net

snapshot

version

vsock

Structs

Body

The Body associated with an HTTP Request or Response.

HttpServer

HTTP Server implementation using Unix Domain Sockets and EPOLL to handle multiple connections on the same thread.

Request

Wrapper over an HTTP Request.

Response

Wrapper over an HTTP Response.

Enums

Method

Supported HTTP Methods.

RequestError

Errors associated with parsing the HTTP Request from a u8 slice.

StatusCode

Wrapper over a response status code.

Version

Supported HTTP Versions.